requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
django-adminlte3==0.1.6
django-bootstrap4==3.0.1
django-localflavor==3.1
djangorestframework==3.12.4
idna==3.2
mysqlclient==2.0.3
Pillow==8.3.1
python-stdnum==1.16
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
